realism
Of course client-side javascript can't scan through the surrounding file structure. The user has to explicitly pick the files, or at least the directory containing them. This also means that there doesn't need to be any special relationship between the location of MusicPlayer.html and the mp3 library - they just both need to be accessible by the browser.